requirements.txt
mo9mo9db==0.1.0
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py>=2.4.0
idna==2.10
multidict==4.7.6
pre-commit==2.13.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-magic==0.4.15
requests==2.25.1
six==1.15.0
tqdm==4.56.0
urllib3==1.26.2
websockets==8.1
yarl==1.5.1
